+++ src/click/core.py
@@ -541,6 +541,9 @@ class Context:
                 info = ctx.to_info_dict()
 
         .. versionadded:: 8.0
+
+        .. versionchanged:: 8.5.1
+            Command dictionaries include the command usage string.
         """
         return {
             "command": self.command.to_info_dict(self),
@@ -1084,6 +1087,7 @@ class Command:
     def to_info_dict(self, ctx: Context) -> dict[str, t.Any]:
         return {
             "name": self.name,
+            "usage": self.get_usage(ctx),
             "params": [param.to_info_dict() for param in self.get_params(ctx)],
             "help": self.help,
             "epilog": self.epilog,